from boolean to two constructors
